Update 116_with_glitch
When you run your code on the actual device each check takes power and also takes times. You want to minimize unnecessary comparisons to maximize remaining power, because the physical devices run on batteries. Thus, make sure 'if' statements have 'else' or 'else if' following them if they wont execute otherwise

You have the right idea with the declaring the variables at top level, in general you don't want to be using any dynamic memory in this, so no pointers-only-memory. I'm not sure if you have to do variable assignment in init though, for example:

float positionA[3] = {800, 800, 0};

should also work, but try that with the compiler you are given. If it gives you wierd warnings about standards try adding the command line arguments --std=c99 or --std=c11 and that should clear it up (and give you better ability to implement googled code).

for counters, use unsigned so you dont get bugs where it goes negative. (unsigned integers go from 0-2^32 and signed integers go from -2^31, 2^31 (or 64 bit or whatever)) 

The static keyword tells the compiler that outside units dont need to access that data, without it you can link another file and say 

extern int counter;

and access the same counter variable here, and as a consequence the compiler has to carry the name of the variable around in the compilation, if you use static when you can the compiler will have an easier time making your code smaller.

on that note when you compile for a competition run and can access command line arguments, make sure to put -O3 or -O2. It makes it hard to debug (so do it on a finished product) but it tells the compiler to go through and do optimizations on your code and make it smaller and use  less power sometimes.

Also, don't forget proper indentation.

You can always reach me at 781-588-2411 or mccue.et@husky.neu.edu (and please do, i'm bored)